+++ src/plugins/console/public/lib/utils/__tests__/utils.test.js
@@ -17,125 +17,80 @@
  * under the License.
  */
 
-const _ = require('lodash');
-const utils = require('../utils');
-const collapsingTests = require('./utils_string_collapsing.txt');
-const expandingTests = require('./utils_string_expanding.txt');
+const utils = require('../');
 
 describe('Utils class', () => {
-  describe('collapseLiteralStrings', () => {
-    it('will collapse multiline strings', () => {
-      const multiline = '{ "foo": """bar\nbaz""" }';
-      expect(utils.collapseLiteralStrings(multiline)).toEqual('{ "foo": "bar\\nbaz" }');
-    });
+  test('extract deprecation messages', function() {
+    expect(
+      utils.extractDeprecationMessages(
+        '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning" "Mon, 27 Feb 2017 14:52:14 GMT"'
+      )
+    ).toEqual(['#! Deprecation: this is a warning']);
+    expect(
+      utils.extractDeprecationMessages(
+        '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning"'
+      )
+    ).toEqual(['#! Deprecation: this is a warning']);
 
-    it('will collapse multiline strings with CRLF endings', () => {
-      const multiline = '{ "foo": """bar\r\nbaz""" }';
-      expect(utils.collapseLiteralStrings(multiline)).toEqual('{ "foo": "bar\\r\\nbaz" }');
-    });
+    expect(
+      utils.extractDeprecationMessages(
+        '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning" "Mon, 27 Feb 2017 14:52:14 GMT", 299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a second warning" "Mon, 27 Feb 2017 14:52:14 GMT"'
+      )
+    ).toEqual(['#! Deprecation: this is a warning', '#! Deprecation: this is a second warning']);
+    expect(
+      utils.extractDeprecationMessages(
+        '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning", 299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a second warning"'
+      )
+    ).toEqual(['#! Deprecation: this is a warning', '#! Deprecation: this is a second warning']);
+
+    expect(
+      utils.extractDeprecationMessages(
+        '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning, and it includes a comma" "Mon, 27 Feb 2017 14:52:14 GMT"'
+      )
+    ).toEqual(['#! Deprecation: this is a warning, and it includes a comma']);
+    expect(
+      utils.extractDeprecationMessages(
+        '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning, and it includes a comma"'
+      )
+    ).toEqual(['#! Deprecation: this is a warning, and it includes a comma']);
+
+    expect(
+      utils.extractDeprecationMessages(
+        '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning, and it includes an escaped backslash \\\\ and a pair of \\"escaped quotes\\"" "Mon, 27 Feb 2017 14:52:14 GMT"'
+      )
+    ).toEqual([
+      '#! Deprecation: this is a warning, and it includes an escaped backslash \\ and a pair of "escaped quotes"',
+    ]);
+    expect(
+      utils.extractDeprecationMessages(
+        '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning, and it includes an escaped backslash \\\\ and a pair of \\"escaped quotes\\""'
+      )
+    ).toEqual([
+      '#! Deprecation: this is a warning, and it includes an escaped backslash \\ and a pair of "escaped quotes"',
+    ]);
   });
 
-  _.each(collapsingTests.split(/^=+$/m), function(fixture) {
-    if (fixture.trim() === '') {
-      return;
-    }
-    fixture = fixture.split(/^-+$/m);
-    const name = fixture[0].trim();
-    const expanded = fixture[1].trim();
-    const collapsed = fixture[2].trim();
-
-    test('Literal collapse - ' + name, function() {
-      expect(utils.collapseLiteralStrings(expanded)).toEqual(collapsed);
-    });
+  test('unescape', function() {
+    expect(utils.unescape('escaped backslash \\\\')).toEqual('escaped backslash \\');
+    expect(utils.unescape('a pair of \\"escaped quotes\\"')).toEqual('a pair of "escaped quotes"');
+    expect(utils.unescape('escaped quotes do not have to come in pairs: \\"')).toEqual(
+      'escaped quotes do not have to come in pairs: "'
+    );
   });
 
-  _.each(expandingTests.split(/^=+$/m), function(fixture) {
-    if (fixture.trim() === '') {
-      return;
-    }
-    fixture = fixture.split(/^-+$/m);
-    const name = fixture[0].trim();
-    const collapsed = fixture[1].trim();
-    const expanded = fixture[2].trim();
-
-    test('Literal expand - ' + name, function() {
-      expect(utils.expandLiteralStrings(collapsed)).toEqual(expanded);
-    });
-
-    test('extract deprecation messages', function() {
-      expect(
-        utils.extractDeprecationMessages(
-          '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning" "Mon, 27 Feb 2017 14:52:14 GMT"'
-        )
-      ).toEqual(['#! Deprecation: this is a warning']);
-      expect(
-        utils.extractDeprecationMessages(
-          '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning"'
-        )
-      ).toEqual(['#! Deprecation: this is a warning']);
-
-      expect(
-        utils.extractDeprecationMessages(
-          '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning" "Mon, 27 Feb 2017 14:52:14 GMT", 299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a second warning" "Mon, 27 Feb 2017 14:52:14 GMT"'
-        )
-      ).toEqual(['#! Deprecation: this is a warning', '#! Deprecation: this is a second warning']);
-      expect(
-        utils.extractDeprecationMessages(
-          '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning", 299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a second warning"'
-        )
-      ).toEqual(['#! Deprecation: this is a warning', '#! Deprecation: this is a second warning']);
-
-      expect(
-        utils.extractDeprecationMessages(
-          '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning, and it includes a comma" "Mon, 27 Feb 2017 14:52:14 GMT"'
-        )
-      ).toEqual(['#! Deprecation: this is a warning, and it includes a comma']);
-      expect(
-        utils.extractDeprecationMessages(
-          '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning, and it includes a comma"'
-        )
-      ).toEqual(['#! Deprecation: this is a warning, and it includes a comma']);
-
-      expect(
-        utils.extractDeprecationMessages(
-          '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning, and it includes an escaped backslash \\\\ and a pair of \\"escaped quotes\\"" "Mon, 27 Feb 2017 14:52:14 GMT"'
-        )
-      ).toEqual([
-        '#! Deprecation: this is a warning, and it includes an escaped backslash \\ and a pair of "escaped quotes"',
-      ]);
-      expect(
-        utils.extractDeprecationMessages(
-          '299 Elasticsearch-6.0.0-alpha1-SNAPSHOT-abcdef1 "this is a warning, and it includes an escaped backslash \\\\ and a pair of \\"escaped quotes\\""'
-        )
-      ).toEqual([
-        '#! Deprecation: this is a warning, and it includes an escaped backslash \\ and a pair of "escaped quotes"',
-      ]);
-    });
-
-    test('unescape', function() {
-      expect(utils.unescape('escaped backslash \\\\')).toEqual('escaped backslash \\');
-      expect(utils.unescape('a pair of \\"escaped quotes\\"')).toEqual(
-        'a pair of "escaped quotes"'
-      );
-      expect(utils.unescape('escaped quotes do not have to come in pairs: \\"')).toEqual(
-        'escaped quotes do not have to come in pairs: "'
-      );
-    });
-
-    test('split on unquoted comma followed by space', function() {
-      expect(utils.splitOnUnquotedCommaSpace('a, b')).toEqual(['a', 'b']);
-      expect(utils.splitOnUnquotedCommaSpace('a,b, c')).toEqual(['a,b', 'c']);
-      expect(utils.splitOnUnquotedCommaSpace('"a, b"')).toEqual(['"a, b"']);
-      expect(utils.splitOnUnquotedCommaSpace('"a, b", c')).toEqual(['"a, b"', 'c']);
-      expect(utils.splitOnUnquotedCommaSpace('"a, b\\", c"')).toEqual(['"a, b\\", c"']);
-      expect(utils.splitOnUnquotedCommaSpace(', a, b')).toEqual(['', 'a', 'b']);
-      expect(utils.splitOnUnquotedCommaSpace('a, b, ')).toEqual(['a', 'b', '']);
-      expect(utils.splitOnUnquotedCommaSpace('\\"a, b", "c, d\\", e", f"')).toEqual([
-        '\\"a',
-        'b", "c',
-        'd\\"',
-        'e", f"',
-      ]);
-    });
+  test('split on unquoted comma followed by space', function() {
+    expect(utils.splitOnUnquotedCommaSpace('a, b')).toEqual(['a', 'b']);
+    expect(utils.splitOnUnquotedCommaSpace('a,b, c')).toEqual(['a,b', 'c']);
+    expect(utils.splitOnUnquotedCommaSpace('"a, b"')).toEqual(['"a, b"']);
+    expect(utils.splitOnUnquotedCommaSpace('"a, b", c')).toEqual(['"a, b"', 'c']);
+    expect(utils.splitOnUnquotedCommaSpace('"a, b\\", c"')).toEqual(['"a, b\\", c"']);
+    expect(utils.splitOnUnquotedCommaSpace(', a, b')).toEqual(['', 'a', 'b']);
+    expect(utils.splitOnUnquotedCommaSpace('a, b, ')).toEqual(['a', 'b', '']);
+    expect(utils.splitOnUnquotedCommaSpace('\\"a, b", "c, d\\", e", f"')).toEqual([
+      '\\"a',
+      'b", "c',
+      'd\\"',
+      'e", f"',
+    ]);
   });
 });
