+++ src/plugins/es_ui_shared/console_lang/ace/modes/index.d.ts
@@ -17,4 +17,16 @@
  * under the License.
  */
 
-export { JsonEditor, OnJsonEditorUpdateHandler } from './components/json_editor';
+import { Editor } from 'brace';
+
+export declare const ElasticsearchSqlHighlightRules: FunctionConstructor;
+export declare const ScriptHighlightRules: FunctionConstructor;
+export declare const XJsonHighlightRules: FunctionConstructor;
+
+export declare const XJsonMode: FunctionConstructor;
+
+/**
+ * @param otherRules Another Ace ruleset
+ * @param embedUnder The state name under which the rules will be embedded. Defaults to "json".
+ */
+export declare const addXJsonToRules: (otherRules: any, embedUnder?: string) => void;
