+++ src/plugins/es_ui_shared/console_lang/ace/modes/lexer_rules/script_highlight_rules.js
