+++ src/plugins/es_ui_shared/console_lang/ace/modes/lexer_rules/x_json_highlight_rules.js
@@ -17,10 +17,15 @@
  * under the License.
  */
 
-const _ = require('lodash');
+import * as _ from 'lodash';
+import ace from 'brace';
+import 'brace/mode/json';
 
 import { ElasticsearchSqlHighlightRules } from './elasticsearch_sql_highlight_rules';
 const { ScriptHighlightRules } = require('./script_highlight_rules');
+
+const { JsonHighlightRules } = ace.acequire('ace/mode/json_highlight_rules');
+const oop = ace.acequire('ace/lib/oop');
 
 const jsonRules = function(root) {
   root = root ? root : 'json';
@@ -145,6 +150,30 @@ const jsonRules = function(root) {
   ];
   return rules;
 };
+
+export function XJsonHighlightRules() {
+  this.$rules = {
+    ...jsonRules('start'),
+  };
+
+  this.embedRules(ScriptHighlightRules, 'script-', [
+    {
+      token: 'punctuation.end_triple_quote',
+      regex: '"""',
+      next: 'pop',
+    },
+  ]);
+
+  this.embedRules(ElasticsearchSqlHighlightRules, 'sql-', [
+    {
+      token: 'punctuation.end_triple_quote',
+      regex: '"""',
+      next: 'pop',
+    },
+  ]);
+}
+
+oop.inherits(XJsonHighlightRules, JsonHighlightRules);
 
 export function addToRules(otherRules, embedUnder) {
   otherRules.$rules = _.defaultsDeep(otherRules.$rules, jsonRules(embedUnder));
