+++ src/plugins/es_ui_shared/console_lang/ace/modes/x_json/index.ts
@@ -17,4 +17,4 @@
  * under the License.
  */
 
-export { JsonEditor, OnJsonEditorUpdateHandler } from './components/json_editor';
+export { XJsonMode } from './x_json_mode';
