+++ src/plugins/es_ui_shared/console_lang/lib/json_xjson_translation_tools/__tests__/json_xjson_translation_tools.test.ts
@@ -1,0 +1,65 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+import _ from 'lodash';
+// @ts-ignore
+import collapsingTests from './utils_string_collapsing.txt';
+// @ts-ignore
+import expandingTests from './utils_string_expanding.txt';
+
+import * as utils from '../index';
+
+describe('JSON to XJSON conversion tools', () => {
+  it('will collapse multiline strings', () => {
+    const multiline = '{ "foo": """bar\nbaz""" }';
+    expect(utils.collapseLiteralStrings(multiline)).toEqual('{ "foo": "bar\\nbaz" }');
+  });
+
+  it('will collapse multiline strings with CRLF endings', () => {
+    const multiline = '{ "foo": """bar\r\nbaz""" }';
+    expect(utils.collapseLiteralStrings(multiline)).toEqual('{ "foo": "bar\\r\\nbaz" }');
+  });
+});
+
+_.each(collapsingTests.split(/^=+$/m), function(fixture) {
+  if (fixture.trim() === '') {
+    return;
+  }
+  fixture = fixture.split(/^-+$/m);
+  const name = fixture[0].trim();
+  const expanded = fixture[1].trim();
+  const collapsed = fixture[2].trim();
+
+  test('Literal collapse - ' + name, function() {
+    expect(utils.collapseLiteralStrings(expanded)).toEqual(collapsed);
+  });
+});
+
+_.each(expandingTests.split(/^=+$/m), function(fixture) {
+  if (fixture.trim() === '') {
+    return;
+  }
+  fixture = fixture.split(/^-+$/m);
+  const name = fixture[0].trim();
+  const collapsed = fixture[1].trim();
+  const expanded = fixture[2].trim();
+
+  test('Literal expand - ' + name, function() {
+    expect(utils.expandLiteralStrings(collapsed)).toEqual(expanded);
+  });
+});
