+++ x-pack/plugins/es_ui_shared/console_lang/ace/modes/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export { installXJsonMode, XJsonMode } from './x_json';
