+++ x-pack/plugins/es_ui_shared/console_lang/ace/modes/x_json/worker/worker.js
