+++ x-pack/plugins/searchprofiler/public/application/containers/main/main.tsx
@@ -5,7 +5,6 @@
  */
 
 import React, { useCallback } from 'react';
-import _ from 'lodash';
 
 import {
   EuiPage,
