+++ x-pack/plugins/searchprofiler/public/application/editor/editor.test.tsx
@@ -6,9 +6,7 @@
 import 'brace';
 import 'brace/mode/json';
 
-jest.mock('./worker', () => {
-  return { workerModule: { id: 'ace/mode/json_worker', src: '' } };
-});
+import '../../../../es_ui_shared/console_lang/mocks';
 
 import { registerTestBed } from '../../../../../test_utils';
 import { Editor, Props } from '.';
