+++ x-pack/plugins/searchprofiler/public/application/utils/check_for_json_errors.ts
@@ -4,12 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-// Convert triple quotes into regular quotes and escape internal quotes.
-function collapseLiteralStrings(data: string) {
-  return data.replace(/"""(?:\s*\r?\n)?((?:.|\r?\n)*?)(?:\r?\n\s*)?"""/g, function(match, literal) {
-    return JSON.stringify(literal);
-  });
-}
+import { collapseLiteralStrings } from '../../../../../../src/plugins/es_ui_shared/console_lang/lib';
 
 export function checkForParseErrors(json: string) {
   const sanitizedJson = collapseLiteralStrings(json);
