+++ x-pack/plugins/watcher/__jest__/client_integration/watch_create_json.test.ts
@@ -3,6 +3,8 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
+import '../../../es_ui_shared/console_lang/mocks';
+
 import { act } from 'react-dom/test-utils';
 import { setupEnvironment, pageHelpers, nextTick, wrapBodyResponse } from './helpers';
 import { WatchCreateJsonTestBed } from './helpers/watch_create_json.helpers';
