+++ x-pack/plugins/watcher/__jest__/client_integration/watch_create_threshold.test.tsx
@@ -3,6 +3,9 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
+
+import '../../../es_ui_shared/console_lang/mocks';
+
 import React from 'react';
 import { act } from 'react-dom/test-utils';
 import axiosXhrAdapter from 'axios/lib/adapters/xhr';
