+++ x-pack/plugins/watcher/__jest__/client_integration/watch_edit.test.ts
@@ -3,6 +3,8 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
+import '../../../es_ui_shared/console_lang/mocks';
+
 import { act } from 'react-dom/test-utils';
 import axiosXhrAdapter from 'axios/lib/adapters/xhr';
 import axios from 'axios';
