+++ x-pack/plugins/watcher/__jest__/client_integration/watch_list.test.ts
@@ -3,6 +3,7 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
+import '../../../es_ui_shared/console_lang/mocks';
 
 import { act } from 'react-dom/test-utils';
 import * as fixtures from '../../test/fixtures';
