+++ x-pack/plugins/watcher/public/application/sections/watch_edit/components/json_watch_edit/json_watch_edit_form.tsx
@@ -20,6 +20,7 @@ import {
 } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n/react';
+
 import { serializeJsonWatch } from '../../../../../../common/lib/serialization';
 import { ErrableFormRow, SectionError, Error as ServerError } from '../../../../components';
 import { onWatchSave } from '../../watch_edit_actions';
@@ -28,6 +29,8 @@ import { goToWatchList } from '../../../../lib/navigation';
 import { RequestFlyout } from '../request_flyout';
 import { useAppContext } from '../../../../app_context';
 
+import { useXJsonMode } from './use_x_json_mode';
+
 export const JsonWatchEditForm = () => {
   const {
     links: { putWatchApiUrl },
@@ -35,6 +38,7 @@ export const JsonWatchEditForm = () => {
   } = useAppContext();
 
   const { watch, setWatchProperty } = useContext(WatchContext);
+  const { xJsonMode, convertToJson, setXJson, xJson } = useXJsonMode(watch.watchString);
 
   const { errors } = watch.validate();
   const hasErrors = !!Object.keys(errors).find(errorKey => errors[errorKey].length >= 1);
@@ -160,9 +164,9 @@ export const JsonWatchEditForm = () => {
           errors={jsonErrors}
         >
           <EuiCodeEditor
-            mode="json"
+            mode={xJsonMode}
             width="100%"
-            theme="github"
+            theme="textmate"
             data-test-subj="jsonEditor"
             aria-label={i18n.translate(
               'xpack.watcher.sections.watchEdit.json.form.watchJsonAriaLabel',
@@ -170,12 +174,14 @@ export const JsonWatchEditForm = () => {
                 defaultMessage: 'Code editor',
               }
             )}
-            value={watch.watchString}
-            onChange={(json: string) => {
+            value={xJson}
+            onChange={(xjson: string) => {
               if (validationError) {
                 setValidationError(null);
               }
-              setWatchProperty('watchString', json);
+              setXJson(xjson);
+              // Keep the watch in sync with the editor content
+              setWatchProperty('watchString', convertToJson(xjson));
             }}
           />
         </ErrableFormRow>
