+++ x-pack/plugins/watcher/public/application/sections/watch_edit/components/threshold_watch_edit/action_fields/webhook_action_fields.tsx
@@ -245,7 +245,7 @@ export const WebhookActionFields: React.FunctionComponent<Props> = ({
           mode="json"
           width="100%"
           height="200px"
-          theme="github"
+          theme="textmate"
           data-test-subj="webhookBodyEditor"
           aria-label={i18n.translate(
             'xpack.watcher.sections.watchEdit.threshold.webhookAction.bodyCodeEditorAriaLabel',
